Split resolver core method signature tests
