+++ tests/docs_truth/repo_hygiene/file_size.rs
@@ -2,5 +2,6 @@ mod core_semantics_splits;
 mod focused_tests;
 mod function_method_template_splits;
 mod lexer_and_monomorphize;
+mod resolver_phase2_splits;
 mod thresholds;
 mod typechecker_test_splits;
